fix: fixed server

+++ src/client/lobby_protocol.cpp
@@ -6,7 +6,7 @@ LobbyProtocol::LobbyProtocol(Socket &a_skt) : was_closed(false), skt(a_skt) {}
 
 uint16_t LobbyProtocol::receive_header() {
   uint16_t header;
-  skt.recvall_bytewise(&header, sizeof(uint8_t), &was_closed);
+  skt.recvall_bytewise(&header, sizeof(uint16_t), &was_closed);
   header = ntohs(header);
   this->skt_was_closed();
   return header;
